+++ sdk/tables/Azure.Data.Tables/CHANGELOG.md
@@ -11,6 +11,7 @@ Thank you to our developer community members who helped to make Azure Tables bet
 ### Added
 
 - Added the `TableErrorCode` type which allows comparison of the `ErrorCode` on `RequestFailedException`s thrown from client operations with a known error value.
+- `TableEntity` and custom entity types now support `BinaryData` properties.
 
 ### Key Bug Fixes